requirements.txt
albumentations==1.3.1
pydantic==1.10.12
wolframalpha~=5.0.0
aiogram~=3.2.0
Pillow
pix2tex
tinydb